Add files via upload
This repository contains the implementation codes for the research paper titled "Intelligent Optimization of Path Control and Accuracy Enhancement of Industrial Robots Using the DE Algorithm." The project focuses on optimizing path control and improving accuracy for a two-degree-of-freedom articulated industrial robot using the Differential Evolution (DE) algorithm. This method is compared with traditional PID and Model Predictive Control (MPC) algorithms to demonstrate superior performance in complex environments.